Update2
The above are the learning notes in a few days, as well as the preliminary completed tank war (Classic) unity project documents, which have not been packaged yet. It is expected that the enemy AI and map generator will be optimized in the future. The content of this update is unity project file